Split out views and update logging formatters
