+++ resource_server_async/endpoints/endpoint.py
@@ -6,17 +6,18 @@ from typing import Any, List, Type
 from django.forms.models import model_to_dict
 from django.utils.text import slugify
 
-from resource_server_async.auth import check_permission as auth_utils_check_permission
 from resource_server_async.cache import get_redis_client
-from resource_server_async.errors import (
+from resource_server_async.rate_limiters import TokenLimiterCheck, TokenRateLimiter
+
+from ..auth import check_permission as auth_utils_check_permission
+from ..errors import (
     BatchUnavailable,
     EndpointNotFound,
     Unauthorized,
 )
-from resource_server_async.models import BatchLog, Endpoint, User
-from resource_server_async.rate_limiters import TokenLimiterCheck, TokenRateLimiter
-from resource_server_async.schemas.batch import BatchSubmit
-from resource_server_async.schemas.endpoints import (
+from ..models import BatchLog, Endpoint, User
+from ..schemas.batch import BatchSubmit
+from ..schemas.endpoints import (
     BatchStatusResult,
     SubmitBatchResult,
     SubmitStreamingTaskResponse,
